# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CodePair API
  version: 0.1.0
  description: |
    Backend API for the CodePair collaborative coding workspace.
    The client creates or joins a session, loads any existing code,
    saves updates per language, and tracks active collaborators.
servers:
  - url: http://localhost:8000
    description: Local FastAPI dev server
tags:
  - name: health
    description: Service status checks
  - name: sessions
    description: Manage collaboration sessions and persisted code
paths:
  /health:
    get:
      summary: Health check
      tags: [health]
      responses:
        "200":
          description: Service is reachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /sessions:
    post:
      summary: Create a new collaboration session
      description: |
        Creates a session that can be shared with collaborators.
        A session ID can be provided (for client-generated IDs) or
        auto-generated by the server. Each session starts with a
        code document per supported language.
      tags: [sessions]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "409":
          description: Session ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      summary: Get session details
      tags: [sessions]
      responses:
        "200":
          description: Session found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions/{sessionId}/code:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      summary: Get saved code for a session
      tags: [sessions]
      parameters:
        - in: query
          name: language
          required: false
          description: Filter to a single language
          schema:
            $ref: "#/components/schemas/Language"
      responses:
        "200":
          description: Latest saved code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeSnapshot"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Save code for a language
      tags: [sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCodeRequest"
      responses:
        "200":
          description: Code saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeDocument"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions/{sessionId}/participants:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      summary: Register a participant in a session
      tags: [sessions]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinSessionRequest"
      responses:
        "200":
          description: Participant registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPresence"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions/{sessionId}/participants/{participantId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/ParticipantId"
    delete:
      summary: Remove a participant from a session
      tags: [sessions]
      responses:
        "200":
          description: Participant removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPresence"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sessions/{sessionId}/history:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      summary: Get code revision history
      tags: [sessions]
      parameters:
        - in: query
          name: language
          required: false
          schema:
            $ref: "#/components/schemas/Language"
          description: Filter to a single language
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Maximum number of revisions to return per language
      responses:
        "200":
          description: Revision history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeHistory"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    SessionId:
      in: path
      name: sessionId
      required: true
      description: Unique session identifier (e.g., abc123)
      schema:
        type: string
        pattern: "^[a-z0-9]{6,12}$"
    ParticipantId:
      in: path
      name: participantId
      required: true
      description: Identifier returned by the join endpoint
      schema:
        type: string
  schemas:
    Language:
      type: string
      enum: [javascript, python, sql]
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
    CreateSessionRequest:
      type: object
      properties:
        sessionId:
          type: string
          description: Optional client-provided session ID. If omitted, the server generates one.
        title:
          type: string
          description: Optional display title for the session
        defaultLanguage:
          $ref: "#/components/schemas/Language"
      additionalProperties: false
    JoinSessionRequest:
      type: object
      properties:
        displayName:
          type: string
          description: Optional participant display name
      additionalProperties: false
    UpdateCodeRequest:
      type: object
      required: [language, content]
      properties:
        language:
          $ref: "#/components/schemas/Language"
        content:
          type: string
          description: Full code content to persist
        author:
          type: string
          description: Optional user identifier for attribution
      additionalProperties: false
    CodeDocument:
      type: object
      properties:
        language:
          $ref: "#/components/schemas/Language"
        content:
          type: string
        version:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          nullable: true
      required: [language, content, version, updatedAt]
    CodeSnapshot:
      type: object
      properties:
        sessionId:
          type: string
        codeByLanguage:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/CodeDocument"
      required: [sessionId, codeByLanguage]
    Participant:
      type: object
      properties:
        participantId:
          type: string
        displayName:
          type: string
          nullable: true
        joinedAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
      required: [participantId, joinedAt, lastSeenAt]
    SessionPresence:
      type: object
      properties:
        sessionId:
          type: string
        participantId:
          type: string
          nullable: true
          description: Present when a participant was newly created
        activeParticipants:
          type: integer
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
      required: [sessionId, activeParticipants, participants]
    Session:
      type: object
      properties:
        sessionId:
          type: string
        title:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
        activeParticipants:
          type: integer
        codeByLanguage:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/CodeDocument"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
      required:
        - sessionId
        - createdAt
        - lastActiveAt
        - activeParticipants
        - codeByLanguage
        - participants
    CodeHistory:
      type: object
      properties:
        sessionId:
          type: string
        revisions:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/CodeDocument"
        limit:
          type: integer
      required: [sessionId, revisions]
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required: [detail]